+++ packages/as-stylus/cli/builder/generate-entrypoint.ts
@@ -7,6 +7,7 @@ const USER_CONTRACT_PATH = "../../contracts/test-1/index.ts";
 
 function generateUserEntrypoint(userFilePath: string): { imports: string; entrypointBody: string } {
   const project = new Project();
+  console.log(userFilePath);
   const sourceFile = project.addSourceFileAtPath(userFilePath);
 
   const functions = sourceFile.getFunctions();
@@ -27,24 +28,22 @@ function generateUserEntrypoint(userFilePath: string): { imports: string; entryp
   };
 }
 
-export function generateEntrypoint() {
-  const __filename = fileURLToPath(import.meta.url);
-  const __dirname = path.dirname(__filename);
+export function generateEntrypoint(targetPath: string): void {
+  const projectRoot = process.cwd(); // base del package ejecutado
+  const indexFilePath = path.join(targetPath, "index.transformed.ts")
+  const { imports, entrypointBody } = generateUserEntrypoint(indexFilePath);
 
-  const userFilePath = path.resolve(__dirname, USER_CONTRACT_PATH);
-  const contractBasePath = path.dirname(userFilePath);
-  const targetPath = path.join(contractBasePath, ".dist");
+  const indexTemplatePath = path.join(projectRoot, "templates", "index.template.ts");
 
-  const { imports, entrypointBody } = generateUserEntrypoint(userFilePath);
+  if (!fs.existsSync(indexTemplatePath)) {
+    console.error(`[as-stylus] Error: index.template.ts not found at ${indexTemplatePath}`);
+    process.exit(1);
+  }
 
-  let indexTemplate = fs.readFileSync(
-    path.resolve(__dirname, "../../templates/index.template.ts"),
-    "utf-8",
-  );
+  let indexTemplate = fs.readFileSync(indexTemplatePath, "utf-8");
 
   indexTemplate = indexTemplate.replace("// @logic_imports", imports);
   indexTemplate = indexTemplate.replace("// @user_entrypoint", entrypointBody);
 
   fs.writeFileSync(path.join(targetPath, "index.ts"), indexTemplate);
-
 }
